requirements.txt
audioread==3.0.1
certifi==2024.12.14
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.12.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.27.0
hyperframe==5.2.0
idna==2.10
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
MarkupSafe==3.0.2
matplotlib==3.10.0
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
networkx==3.4.2
noisereduce==3.0.3
numba==0.60.0
numpy==2.0.2
openai-whisper @ git+https://github.com/openai/whisper.git@517a43ecd132a2089d85f4ebc044728a71d49f6e
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
pycparser==2.22
pydub==0.25.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rfc3986==1.5.0
safetensors==0.5.0
scikit-learn==1.6.0
scipy==1.15.0
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
soundfile==0.13.0
soxr==0.5.0.post1
sympy==1.13.3
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
torch==2.2.2
tqdm==4.67.1
transformers==4.47.1
typing_extensions==4.12.2
urllib3==2.3.0
